docs(pipeline-status-ledger): enforce full concept inventory, fail on missing file
Fixes 4 more Codex review findings (3xP1, 1xP2) on PR #1019:

- load_pipeline_status() now errors (not silently skips) when the
  ledger file is deleted/renamed -- previously the only 'is this
  ledger real' signal, its absence, produced zero findings.
- check_pipeline_status_ledger() now checks the 'concepts' mapping's
  keys against an explicit _PIPELINE_REQUIRED_CONCEPTS set in both
  directions (missing and unknown/misspelled), not just validating
  whichever entries happen to be present -- a dropped or typo'd
  concept key previously produced zero findings as long as one valid
  row remained.
- ADR-063's pointer to the ledger no longer says 'generated from the
  plan's own state' -- it's hand-maintained content behind a
  structural validator, not an automatically-derived summary; the
  wording implied a generator that doesn't exist.
- plan.md's mechanical definition-of-done item (3) referenced an
  undefined 'SemanticSnapshot' type; fixed to name the real type
  (SemanticIR, ADR-063 Phase 6's own primitive) and the real persisted
  container (AbiSnapshot.semantic_ir), and sharpened item (4) to name
  the actual legacy collections a migrated checker must stop reading.

Adding the first two fixes inline pushed check_docs_contract.py over
the AI-readiness file-size hard cap (2000 lines), so this also splits
the whole pipeline-status-ledger check into a new leaf module,
scripts/pipeline_status_ledger.py, the same way adr_status_sync.py/
engine_cli_boundary.py were split out of check_ai_readiness.py for the
identical reason -- a mechanical extraction (unchanged logic), not a
redesign. Added its scripts/CLAUDE.md inventory row per that file's
own 'Adding a new script' convention.

Verified via check_docs_contract.py, adr_status_sync.py,
check_ai_readiness.py (0 errors, back to the original 137 warnings —
the file-size WARN moved with the split, no new warning), ruff
check/format, mypy (no new errors beyond one pre-existing, unrelated
one in an untouched function), tests/test_docs_contract.py +
tests/test_ai_readiness.py (218 passed), mkdocs build --strict, and
ad hoc scripts confirming both new checks actually fire on malformed
input.
